tclibc-musl: use musl's virtual/gettext implementation
musl implements the GNU gettext API version 1 and 2
  http://www.openwall.com/lists/musl/2015/04/16/3

Generally, there should be no need need for the
actual GNU/gettext in systems using musl, so with
this change systems using musl can avoid a needless
dependency on a GPLv3+ tool.

If an application ever appears that needs
GNU/gettext API version 3, we can think about a way
to support it at that time.

Signed-off-by: André Draszik <adraszik@tycoint.com>
Acked-by: Sylvain Lemieux <slemieux@tycoint.com>